requirements.txt
annotated-types
anyio
cachetools
certifi
charset-normalizer
contourpy
cycler
einops
exceptiongroup
filelock
fonttools
fsspec
google-auth
google-genai
h11
hf-xet
httpcore
httpx
huggingface-hub
idna
imageio
importlib-metadata
importlib-resources
jinja2
joblib
kiwisolver
lazy-loader
markupsafe
matplotlib
mpmath
networkx
numpy
nvidia-cublas-cu12
nvidia-cuda-cupti-cu12
nvidia-cuda-nvrtc-cu12
nvidia-cuda-runtime-cu12
nvidia-cudnn-cu12
nvidia-cufft-cu12
nvidia-cufile-cu12
nvidia-curand-cu12
nvidia-cusolver-cu12
nvidia-cusparse-cu12
nvidia-cusparselt-cu12
nvidia-nccl-cu12
nvidia-nvjitlink-cu12
nvidia-nvtx-cu12
opencv-python
packaging
pillow
pyasn1
pyasn1-modules
pydantic
pydantic-core
pyparsing
python-dateutil
python-dotenv
pyyaml
regex
requests
rsa
safetensors
scikit-image
scikit-learn
scipy
sentence-transformers
six
sympy
tenacity
threadpoolctl
tifffile
tokenizers
torch
torchvision
tqdm
transformers
triton
typing-extensions
typing-inspection
urllib3
websockets
zipp
